Fixed compilation errors + User service tests

+++ api/src/main/scala/com/notnamed/api/database/entity/User.scala
@@ -3,4 +3,4 @@ package com.notnamed.api.database.entity
 import com.notnamed.commons.database.entity.KeyedEntity
 
 
-case class User(id: Option[Int], email: String,password: String) extends KeyedEntity
+case class User(id: Option[Long], email: String,password: String) extends KeyedEntity
